requirements.txt
# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.10.0
langchain-google-genai==0.0.6

# Web and API
requests==2.31.0

# Streamlit UI
streamlit==1.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0

